third_party/WebKit/LayoutTests/platform/win/virtual/mojo-loading/http/tests/cookies/same-site/popup-same-site-expected.txt

This is a testharness.js-based test.
PASS 'example.test' is same-site with itself, so samesite cookies are sent. 
Harness: the test ran to completion.
